fix: separate client product API
