+++ components/products/create-product-form.tsx
@@ -4,7 +4,8 @@ import type React from "react"
 
 import { useState } from "react"
 import { useRouter } from "next/navigation"
-import { ProductsAPI, type CreateProductData } from "@/lib/api/products"
+import { ProductsAPI } from "@/lib/api/products-client"
+import type { CreateProductData } from "@/lib/api/products"
 import type { Supplier } from "@/lib/api/suppliers"
 import { Button } from "@/components/ui/button"
 import { Input } from "@/components/ui/input"
